ob-cli: logic of CLI extracted in a separate file.
extracted main cli logic into main.py file.
